backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pymongo==4.6.1
motor==3.3.2
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyMuPDF==1.23.8
pdfplumber==0.10.3
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3
pandas==2.1.4
spacy==3.7.2
keybert==0.8.4
sentence-transformers==2.2.2
scikit-learn==1.4.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai>=0.4.0
